Do not show tutorial indicator in wallpaper preview
Add an isPreview option in the indicator view binder so that when it is
true, the view model never shows the indicator.

This change also sets the visibility to GONE by default when applying
constraints in the tutorial indicator section because even though the
view itself is default to GONE, it is reverted to VISIBLE when applying
constraints, and then back to GONE once the first signal from the binder
comes through, causing a quick flash when the device boots up.

Test: verify there is no flash of the indicator after rebooting
Test: verify indicator not shown in the wallpaper preview
Test: verify indicator does show when tutorial is available
Bug: 320534166
Fix: 320534166
Flag: ACONFIG com.android.systemui.communal_hub DEVELOPMENT
Change-Id: I2d432edd048ef46df3ae351a14ae5d2650aa1704