[flexiglass] Fixes issue where bouncer exited upon orientation change.
Previous logic was set up to leave the bouncer scene when the IME
(software keyboard) became hidden. Unfortunately, this was the wrong
signal as it would happen also when changing device configuration (for
example, when changing device orientation).

A better (and less complex) approach is shown in this CL, we listen for
the back navigation keyevent from the IME.

Fix: 322518343
Test: manually verified that changing orientations doesn't dismiss the
bouncer
Test: unit tests updated
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT

Change-Id: I53b6c2c605d3621ce78247305cbe3a900d2ca3fd