Adding UserProperty for Restricting PS Items on home screen.
This property will be used to restrict the addition of items
such as such as Apps Pending Installation, Widgets, Custom App
Shortcuts, etc. on Launcher home screen.

Primary use case if for users/profiles which have restricted
visibility in all surfaces due to product definition (such as
private profile) for which icons placed on home screen can be
a giveaway for the the presence of a separate app installation
space.
For such profiles, this property can be set as true (default value
is false), and accordingly framework APIs will restrict
information about home screen items to Launcher(s).

Bug: 287975131
Test: atest UserManagerServiceUserPropertiesTest
Test: atest UserManagerServiceUserTypeTest
Test: atest FrameworksServicesTests:UserManagerTest
Change-Id: Iab0805bd92d43bf31cc12ae17f748800d0c40139