Fill in policies for default zen rules.
With modes_api, all zen rules must have fully-filled-in policies. However, the default XML config file doesn't include the default policy (which would be the same as the overall config) for the default system rules. This change fills those policies in so they are fully specified when ZenModeHelper is created.

Fixes: 322689667
Bug: 312739308
Test: repro manually on freshly wiped device before fix; confirmed it works correctly and does not crash after fix.
Test: ZenModeHelperTest
Change-Id: Ic156d997aa4ac44374726ef5ce283a41f75139ea