Use CompositingStrategy.ModulateAlpha when fading elements
This CL switches the compositionStrategy of elements to ModulateAlpha
when fading the elements. While this will not always lead to the most
correct result visually if the UI has overlapping rendering (which it
should not in SysUI case), it leads to much better performance.

In a follow-up CL, I will add a way to change the compositionStrategy
for cases where there is overlapping rendering.

Bug: 291071158
Test: PlatformComposeSceneTransitionLayoutTests
Flag: N/A
Change-Id: I96a4457cba36d0fbd1d15861ea1571e74cfa7d77