[PreVerfiedDomains] shell command to report domain verification agent
We need a way for CTS to disable the current domain verification agent
on device. The disablement is needed to test the pre-verified states.
Otherwise, the pre-verified states will get overwritten by the domain
verification agent very soon after the app installation. This CL add a
shell command that reports the component name of the current domain
verification agent.

BUG: 307327678
Test: adb shell pm get-domain-verification-agent

Change-Id: I9ae4113ed22222674b84c79e29a2fe957f54a991