Implement auto-enter PiP2 in gesture nav [1/N]
Implement auto-enter PiP in PiP2 while in gesture
navigation mode.
Launcher handles the animation as a part of the first
TO_FRONT transition (no changes there).
For now only startSwipePipToHome() and stopSwipePipToHome()
IPC calls are set up to let launcher start PiP animation.
The plan is to add the pip animation listener to properly
update Launcher state in [2/N] CL of this series (similar to pip1).

This is followed by a call to requestStartTransition()
-> PipTransition#handleRequest() for TRANSIT_PIP type of a transition.

The flow here overlaps with other enter PiP2 flows, as we use
WindowContainerTransaction#movePipActivityToPinnedRootTask()
to update the WM state with the new bounds (unlike in PiP1).

Then, we recognize that this is a swipe PiP to Home animation
in PipTransition#startAnimation() and we carry out a no-op animation
essentially finishing TRANSIT_PIP. The plan is to continue overlay
animation here as well in later CLs of this series. So overlay is
not animating properly for now - hence testing would require non-null
srcRechHint in PictureInPictureParams.

Bug: 325481148
Test: swipe auto-enter activity (srcRectHint != null) to PiP
Change-Id: If171678181b8b9e0006c74fe41cd441ec25a10b4