Get person avatar icon on a bg thread
We're making a binder call to launcher, and it shouldn't block the main
thread.

Tracking this change and the other changes in the child bugs of b/315143361
under the same generic flag.

Fix: 320274812
Bug: 315143361
Test: posted a few high prio conversation notifs and checked that they
look good on aod, the shelf and status bar
Flag: ACONFIG com.android.systemui.notifications_background_icons DEVELOPMENT

Change-Id: Ie03cf3698c2d46c75012792a5628a5b1068b244b