Enter PiP with fixed rotation restore reentry state correctly
Previously the maxSize was calculated on the previous orientation (e.g.
landscape), and thus lead to incorrect destination bounds.
This change makes sure that the maxSize used is for the correct
rotation and that PipTransition also applies the scaling correctly.

Also refactoring some of the PipTouchHandler to avoid duplicate
calculations.

Bug: 326490788
Test: after fix http://recall/-/ekEuGtt9d9HWqkUtAzpHx8/bngzAYINo6LsYv8Tdt9y31
Flag: none

Change-Id: I1e10b7d67f49399d9b869d02cff923b7ec104d88