Remove unused references to android.security.KeyStore
Many files that reference android.security.KeyStore don't actually do
anything with it.  This is apparently because the functionality of this
class got moved to other places, but it was forgotten to remove the
actual references to the class.  This CL removes all such references
from frameworks/base.  Where made possible by removing an import of
android.security.KeyStore, it also updates references to
java.security.KeyStore to be simply KeyStore.  java.security.KeyStore is
the real KeyStore that is part of the public API.

Bug: 326508120
Test: presubmit
Change-Id: Ie63b8dfc8bb562b33e6e529d537bde64ae2c6d53