Expose sim bouncer messages as events instead of state.
There is no need to store state in the repository, all the messages are in response to user events
The default message is derived from the already stored state.

The subsequent CLs change the view and viewModel layer to use this SharedFlow.

Bug: 299343534
Flag: ACONFIG com.android.systemui.compose_bouncer DEVELOPMENT
Test: atest SimBouncerInteractorTest
Change-Id: Ic86aebc99594d6ed3c2779b0a8e862764c3220f0