Deduplicate & group entries based on a designated dedup id (entryGroupId)
Before, we deduplication entries based on password / passkey usernames
and unknown typed credential title. However, it is possible that the unknown
typed credential will want to customize the value used for
deduplication, a value that will be different from the entry title it
want to display. This change allows that.

Bug: 306722792
Test: Screenshots attached to bug
Change-Id: I854e7afca7e999493437736d29e9ab866dfafd7a