Move createChooser behavior change behind a flag
ag/26138712 added the new extra (and the rest of the API) but changed
the createChooser() behavior without regard to the flag. This caused
some test breakage and created some complexity with respect to
pre-V CTS tests checking for something that wasn't part of android U.

This change puts the createChooser behavior change behind the flag with
the rest of the API (and the corresponding CTS test honors that change
in this topic).

The IntentResolver (which receives this Intent) already honors both
extras so no change is necessary there.

Bug: 325707235
Test: atest IntentTest with the flag on and off
Test: Manual testing with the ShareTest app
Change-Id: I43d937722d600c5f1e3e2ddeb99b05beeab81621