Fix notification scrim missing when opening shade over glanceable hub
Before this change, the scrims were entirely missing when the
notification shade was opened over the hub. The previous change
69a2dd7558703dfb47c2657a9e5ce5650b89dff5 neglected the fact that the
scrims behave differently when dreaming vs on the lockscreen.

This change adds a new ScrimState for when the hub is open over the
dream and uses the same logic as when opening the shade/bouncer over
the dream to set the scrim alphas and tints.

Bug: 323036128
Fixed: 323036128
Test: atest ScrimControllerTest CentralSurfacesImplTest
      also manually verified on device
Flag: ACONFIG com.android.systemui.communal_hub TEAMFOOD
Change-Id: Iec1443357eba9c4c2bd8a2e386120ffb12fed8b2