Build an isHeadsUpAnimatingAway signal from the NSSL
The logic for updating the isHeadsUpAnimatingAway state was spread out
between NPVC and HeadsUpManagerPhone. However this is a view state, that
is controlled by the NSSL animations. This CL moves managing
isHeadsUpAnimatingAway to the NSSL, and pipes it to a repository to be
able to share it with others, even after NPVC is shut down.

Bug: 328390331
Test: observe if HUNs are visible during their outro animation
Test: atest NotificationStackScrollLayoutTest NotificationPanelViewControllerTest NotificationStackScrollLayoutTest
Flag: ACONFIG com.android.systemui.notifications_heads_up_refactor DEVELOPMENT
Change-Id: I704641a041758695cd95767b10c4e55886e4504f